[53] refactor: move Affiliation structs and implementations to a separate file
